Add src/main/resources as a resource location
Usually this is the default but since we declare the LICENCE and NOTICE
files here, the default gets overridden.
Currently this is not an issue, but to avoid a surprise in the future in
case someone expects the conventional resources path to work, I am
adding the default path back in.